Separate BlockChain and Discovery service tests from PeerListenerTest
